+++ src/core_modules/capture-core/components/DataEntry/dataEntryOutput/withFeedbackOutput.js
@@ -1,95 +1,35 @@
 // @flow
 import * as React from 'react';
 import { connect } from 'react-redux';
-import Card from '@material-ui/core/Card';
-import { Menu, MenuItem } from '@dhis2/ui';
-import { withStyles } from '@material-ui/core/styles';
 import i18n from '@dhis2/d2-i18n';
 import { getDataEntryKey } from '../common/getDataEntryKey';
 import { withDataEntryOutput } from './withDataEntryOutput';
-
+import { WidgetFeedback } from '../../WidgetFeedback';
+import type { FilteredText, FilteredKeyValue } from '../../WidgetFeedback';
 
 type Props = {
     feedbackItems: {
-        displayTexts: [{ key: string, value: string}],
-        displayKeyValuePairs: [{ key: string, value: string}],
-    },
-    classes: {
-        listItem: string,
-        card: string,
-        keyValuePairKey: string,
+        displayTexts: Array<FilteredText>,
+        displayKeyValuePairs: Array<FilteredKeyValue>,
     },
 };
 
-const styles = (theme: Theme) => ({
-    listItem: {
-        display: 'flex',
-        backgroundColor: '#f5f5f5 !important',
-        paddingLeft: theme.typography.pxToRem(10),
-        marginTop: theme.typography.pxToRem(8),
-    },
-    keyValuePairKey: {
-        flexGrow: 1,
-        margin: 0,
-    },
-    keyValue: {
-        margin: 0,
-        fontSize: '0.875rem',
-    },
-    card: {
-        padding: theme.typography.pxToRem(10),
-        borderRadius: theme.typography.pxToRem(5),
-    },
-    labelContainer: {
-        display: 'flex',
-    },
-});
-
 const getFeedbackOutput = () =>
     class FeedbackOutputBuilder extends React.Component<Props> {
-        renderFeedbackItems = (feedbackItems: any, classes: any) =>
-            (<div>
-                {feedbackItems.displayTexts &&
-                    feedbackItems.displayTexts.map(item => (
-                        <MenuItem
-                            dense
-                            key={item.id}
-                            className={classes.listItem}
-                            button={false}
-                            label={<p className={classes.keyValuePairKey}> {item.message} </p>}
-                        />
-                    ),
-                    )}
-                {feedbackItems.displayKeyValuePairs &&
-                    feedbackItems.displayKeyValuePairs.map(item => (
-                        <MenuItem
-                            key={item.id}
-                            button={false}
-                            className={classes.listItem}
-                            dense
-                            label={
-                                <div className={classes.labelContainer}>
-                                    <p className={classes.keyValuePairKey}> {item.key} </p>
-                                    <p className={classes.keyValue}> {item.value} </p>
-                                </div>
-                            }
-                        />
-                    ),
-                    )}
-            </div>)
+        getItems = () => {
+            const { feedbackItems } = this.props;
+            const displayTexts = feedbackItems?.displayTexts || [];
+            const displayKeyValuePairs = feedbackItems?.displayKeyValuePairs || [];
+            return [...displayTexts, ...displayKeyValuePairs];
+        }
 
         render = () => {
-            const { feedbackItems, classes } = this.props;
-            const hasItems = feedbackItems && (feedbackItems.displayTexts || feedbackItems.displayKeyValuePairs);
+            const feedback = this.getItems();
+            const hasItems = feedback.length > 0;
             return (
                 <div>
                     {hasItems &&
-                        <Card className={classes.card}>
-                            {i18n.t('Feedback')}
-                            <Menu dense>
-                                {feedbackItems && this.renderFeedbackItems(feedbackItems, classes)}
-                            </Menu>
-                        </Card>
+                        <WidgetFeedback feedback={feedback} emptyText={i18n.t('No feedback for this event yet')} />
                     }
                 </div>
             );
@@ -113,4 +53,4 @@ export const withFeedbackOutput = () =>
 
         withDataEntryOutput()(
             InnerComponent,
-            withStyles(styles)(connect(mapStateToProps, mapDispatchToProps)(getFeedbackOutput())));
+            connect(mapStateToProps, mapDispatchToProps)(getFeedbackOutput()));