+++ src/core_modules/capture-core/components/Pages/common/WidgetEventEditWrapper/WidgetEventEditWrapper.js
@@ -5,6 +5,7 @@ import { pageStatuses } from '../../EnrollmentEditEvent/EnrollmentEditEventPage.
 import { IncompleteSelectionsMessage } from '../../../IncompleteSelectionsMessage';
 import { WidgetEventEdit } from '../../../WidgetEventEdit';
 import type { Props } from '../../../WidgetEventEdit/widgetEventEdit.types';
+import { useMetadataForProgramStage } from '../../../DataEntries/common/ProgramStage/useMetadataForProgramStage';
 
 type WidgetProps = {|
     pageStatus: string,
@@ -12,6 +13,18 @@ type WidgetProps = {|
 |}
 
 export const WidgetEventEditWrapper = ({ pageStatus, ...passOnProps }: WidgetProps) => {
+    const {
+        programId,
+        stageId,
+    } = passOnProps;
+
+    const {
+        formFoundation,
+        stage,
+        isLoading,
+        isError,
+    } = useMetadataForProgramStage({ programId, stageId });
+
     if (pageStatus === pageStatuses.WITHOUT_ORG_UNIT_SELECTED) {
         return (
             <IncompleteSelectionsMessage>
@@ -26,9 +39,29 @@ export const WidgetEventEditWrapper = ({ pageStatus, ...passOnProps }: WidgetPro
         );
     }
 
+    if (isLoading) {
+        return (
+            <div>
+                {i18n.t('Loading')}
+            </div>
+        );
+    }
+
+    if (!formFoundation || !stage || isError) {
+        return (
+            <div>
+                {i18n.t('An error occurred while loading the form')}
+            </div>
+        );
+    }
+
     return (
         <WidgetEventEdit
             {...passOnProps}
+            stage={stage}
+            formFoundation={formFoundation}
+            programId={programId}
+            stageId={stageId}
         />
     );
 };