+++ src/core_modules/capture-core/metaDataMemoryStoreBuilders/programs/factory/enrollment/EnrollmentFactory.js
@@ -18,7 +18,7 @@ import { DataElementFactory } from './DataElementFactory';
 import type { ConstructorInput } from './enrollmentFactory.types';
 import { transformTrackerNode } from '../transformNodeFuntions/transformNodeFunctions';
 import { FormFieldPluginConfig } from '../../../../metaData/FormFieldPluginConfig';
-import type { DataEntryFormConfig } from '../../../../components/DataEntries/common/TEIAndEnrollment/useMetadataForRegistrationForm/types';
+import type { DataEntryFormConfig } from '../../../../components/DataEntries/common/TEIAndEnrollment';
 import { FormFieldTypes } from '../../../../components/D2Form/FormFieldPlugin/FormFieldPlugin.const';
 import {
     FieldElementObjectTypes,
@@ -181,6 +181,7 @@ export class EnrollmentFactory {
         cachedProgramTrackedEntityAttributes?: Array<CachedProgramTrackedEntityAttribute>,
         cachedSectionCustomLabel: string,
         cachedSectionCustomId: string,
+        description: string,
     ) {
         if (!cachedProgramTrackedEntityAttributes?.length) {
             return null;
@@ -189,6 +190,7 @@ export class EnrollmentFactory {
         const section = new Section((o) => {
             o.id = cachedSectionCustomId;
             o.name = cachedSectionCustomLabel;
+            o.displayDescription = description;
             o.group = Section.groups.ENROLLMENT;
         });
 
@@ -302,6 +304,7 @@ export class EnrollmentFactory {
                             attributes,
                             formConfigSection.name ?? sectionMetadata?.displayFormName ?? i18n.t('Profile'),
                             formConfigSection.id,
+                            sectionMetadata?.displayDescription ?? '',
                         );
                         section && enrollmentForm.addSection(section);
                     });
@@ -312,6 +315,7 @@ export class EnrollmentFactory {
                             programSection.trackedEntityAttributes.map(id => trackedEntityAttributeDictionary[id]),
                             programSection.displayFormName,
                             programSection.id,
+                            programSection.displayDescription,
                         );
                         section && enrollmentForm.addSection(section);
                     });